Former-commit-id: 8ccb7bb23e63208a2b517bf1913f93e794c6294e

+++ Backend/PIGGY/src/main/java/com/piggy/PIGGY/controller/PostRestController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Required;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -106,15 +107,17 @@ public class PostRestController {
 	public ResponseEntity<Object> update(@PathVariable Long pId, 
 			@RequestParam String content,
 			@RequestParam Integer isLike,
-			@RequestParam("file") MultipartFile file){
+			@RequestParam Boolean visited,
+			@RequestParam(value="file", required=false) MultipartFile file){
 		try {
 			log.trace("PostRestController - update");
 			Post post = pService.findById(pId);
-			PostInputDto dto = new PostInputDto(post.getStore().getSId(), content, post.getVisited(), isLike);
+			PostInputDto dto = new PostInputDto(post.getStore().getSId(), content, visited, isLike);
 			post = pService.update(pId, dto);
-			
-			if(!file.getName().equals(post.getImageName())) {
+			if(file != null && post.getImageName()!=null) {
 				fileService.deleteImage(post.getImageName());
+			}
+			if(file != null && !file.getName().equals(post.getImageName())) {
 				Map<String, Object> responseImage = fileService.uploadImage(file, "post");
 				PostImageDto imageDto = new PostImageDto(pId, responseImage.get("image").toString(), responseImage.get("imageName").toString());
 				post = pService.updateImage(pId, imageDto);
@@ -132,7 +135,9 @@ public class PostRestController {
 		try {
 			log.trace("PostRestController - delete", pId);
 			Post post = pService.findById(pId);
-			fileService.deleteImage(post.getImageName());
+			if(post.getImageName()!=null) {
+				fileService.deleteImage(post.getImageName());
+			}
 			pService.delete(pId);
 			return new ResponseEntity<Object>(new ResultDto(true, 1, "삭제되었습니다."), HttpStatus.OK);
 		} catch (Exception e) {
