+++ Backend/PIGGY/src/main/java/com/piggy/PIGGY/controller/UserRestController.java
@@ -22,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.piggy.PIGGY.dto.ResultDto;
+import com.piggy.PIGGY.dto.SigninDto;
 import com.piggy.PIGGY.dto.SignupDto;
 import com.piggy.PIGGY.dto.UserDto;
 import com.piggy.PIGGY.entity.User;
@@ -105,12 +106,12 @@ public class UserRestController {
 		@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
 	@ApiOperation(value = "비밀번호 확인")
 	@PostMapping("/checkPassword")
-	public ResponseEntity<Object> checkPassword(@RequestBody String password) {
+	public ResponseEntity<Object> checkPassword(@RequestBody SigninDto dto) {
 		try {
 			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 			String uId = authentication.getName();
 			
-			if(uService.checkPassword(Long.parseLong(uId), password))
+			if(uService.checkPassword(Long.parseLong(uId), dto.getPassword()))
 				return new ResponseEntity<Object>(new ResultDto(true, 1, "비밀번호가 확인되었습니다."), HttpStatus.OK);
 			else
 				return new ResponseEntity<Object>(new ResultDto(false, -1, "비밀번호가 틀렸습니다"), HttpStatus.ACCEPTED);
@@ -121,13 +122,13 @@ public class UserRestController {
 	
 	@ApiOperation(value = "비밀번호수정")
 	@PutMapping("/updatePassword")
-	public ResponseEntity<Object> updatePassword(@RequestParam String email, @RequestBody String password) {
+	public ResponseEntity<Object> updatePassword(@RequestBody SigninDto dto) {
 		try {
-			User user = uService.findByEmail(email);
+			User user = uService.findByEmail(dto.getEmail());
 			if (!user.getEmailCertify().equals("Y"))
 				return new ResponseEntity<Object>(new ResultDto(false, -1, "이메일 인증이 되지 않은 유저"), HttpStatus.OK);
 			
-			user = uService.updatePassword(user.getUId(), password);
+			user = uService.updatePassword(user.getUId(), dto.getPassword());
 			UserDto output = MapperUtils.map(user, UserDto.class);
 			return new ResponseEntity<Object>(output, HttpStatus.OK);
 		} catch (Exception e) {
@@ -149,17 +150,20 @@ public class UserRestController {
 	
 	@ApiOperation(value = "해당 유저 이미지 수정")
 	@PostMapping("/uploadImage/{uId}")
-	public ResponseEntity<Object> postImage(@PathVariable Long uId , @RequestParam("file") MultipartFile file) {
+	public ResponseEntity<Object> uploadImage(@PathVariable Long uId , @RequestParam(value="file", required=false) MultipartFile file) {
 		try {
-			Map<String, Object> responseImage = fileService.uploadImage(file, "user");
 			User origin = uService.findById(uId);
-			String originImage = origin.getImageName();
-			String newImageName = responseImage.get("imageName").toString();
-			if(originImage != null && !originImage.equals(newImageName)) {
-				fileService.deleteImage(originImage);
+			UserDto output = MapperUtils.map(origin, UserDto.class);
+			if(file != null) {
+				Map<String, Object> responseImage = fileService.uploadImage(file, "user");
+				String originImage = origin.getImageName();
+				String newImageName = responseImage.get("imageName").toString();
+				if(originImage != null && !originImage.equals(newImageName)) {
+					fileService.deleteImage(originImage);
+				}
+				User user = uService.updateImage(uId, responseImage.get("image").toString(), newImageName);
+				output = MapperUtils.map(user, UserDto.class);
 			}
-			User user = uService.updateImage(uId, responseImage.get("image").toString(), newImageName);
-			UserDto output = MapperUtils.map(user, UserDto.class);
 			return new ResponseEntity<Object>(output, HttpStatus.OK);
 		} catch (Exception e) {
 			return new ResponseEntity<Object>(e.getMessage(), HttpStatus.OK);
